# MyYtlCar/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API's",
    "description": "Connected Car Demo",
    "version": "1.0"
  },
  "produces": ["application/json"],  
  "host": "MyYtlCar.mybluemix.net",
  "basePath": "/",
  "paths": {
...
	 "/api/v1/car/pair": {
		  "post": {
			"x-swagger-router-controller": "car",
			"operationId": "carpair",
			"tags": ["Car"],
			"description": "Pair the car device with user.",   
			"consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
			"parameters": [
			  {
				"name": "access_token",
				"in": "header",
				"description": "Access token",
				"required": true,
				"type": "string"
			  },  
			  {
				"name": "body",
                "in": "body",               
                "description": "car pair needs the info of car",
                "required": true,
                "schema": {
                            "$ref": "#/definitions/carpair"
                        }
              }
			],		
			"responses": {
			  "default": {
				"description": "Invalid request.",
				"schema": {
				  "$ref": "#/definitions/Error"
				}
			  },
			  "200": {
				"description": "Successful request.",
				"schema": {
				    "$ref": "#/definitions/Success"
				}
			  }
			}
		  }
	  },
...
  },
  "definitions": {
...
	 "carpair": {
            "type": "object",
            "properties": {
                
                "deviceid": {
                    "type": "string"
                },
                "devicetype": {
                    "type": "string"
                },
                "devicetoken": {
                    "type": "string"
                },
                "orgid": {
                    "type": "string"
                },
                "carname": {
                    "type": "string"
                }
            }
        },
	"Error": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
	"Success": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    }
  }
}